refactor(types): share domain types across indexer and graphql-api
packages/types exported the platform's domain types, but services/indexer
imported nothing from it and services/graphql-api imported only
CampaignStatus, redefining every other domain shape locally. An on-chain
event was therefore described by three independent, hand-maintained type
definitions that could silently drift apart.

Move the shared definitions into packages/types and have both services
import them:

- packages/types/src/indexer.ts holds IndexerEvent, which the indexer now
  re-exports from rpc-client.ts so its own import paths are unchanged.
- packages/types/src/contract.ts holds the raw Soroban view shapes, with
  `status` typed as the canonical CampaignStatus instead of a second,
  identical union.
- packages/types/src/graphql-server.ts holds the resolver-side view of the
  GraphQL schema (BigInt scalars as bigint), alongside the existing
  client-side codegen types in graphql.ts.

services/graphql-api/src/types.ts keeps only the three types that depend on
the service's own runtime (Context, DataLoaders, AuthenticatedUser) and
re-exports the rest, so a change in packages/types is now a compile error in
both services.

A GraphQLFieldParity drift guard in graphql-server.ts fails to compile if
the server and client views of the schema gain or lose a field
independently, and the new test covers the enums, which survive to runtime.

Closes #1055
