+++ services/indexer/package.json
@@ -14,6 +14,7 @@
   },
   "dependencies": {
     "@fund-my-cause/shared-utils": "file:../../packages/shared-utils",
+    "@fund-my-cause/types": "file:../../packages/types",
     "express": "^4.18.2",
     "@stellar/stellar-sdk": "^12.0.0",
     "pino": "^8.17.2",
